Added the volume checking model and the ability to decide models in the pipeline config. The pipeline config now needs to have a "model" parameter for the scan pages, and the options are "VolumeCheckModel" and "RackScanModel".
The volume check page loads the csv file and verifies that the number of tubes match the input amount, and then has the capability to save to tube rack when the user clicks to save the volume quantities.

This is still a bit rough, so further improvements can be made.